Move snap-sync structures to domains.
