+++ crates/subspace-malicious-operator/src/malicious_domain_instance_starter.rs
@@ -1,6 +1,7 @@
 use crate::malicious_bundle_producer::MaliciousBundleProducer;
 use crate::{create_malicious_operator_configuration, DomainCli};
 use cross_domain_message_gossip::{ChainMsg, Message};
+use domain_client_operator::snap_sync::ConsensusChainSyncParams;
 use domain_client_operator::{BootstrapResult, OperatorStreams};
 use domain_eth_service::provider::EthProvider;
 use domain_eth_service::DefaultEthConfig;
@@ -28,7 +29,6 @@ use std::sync::Arc;
 use subspace_runtime::RuntimeApi as CRuntimeApi;
 use subspace_runtime_primitives::opaque::Block as CBlock;
 use subspace_runtime_primitives::AccountId;
-use subspace_service::domains::ConsensusChainSyncParams;
 use subspace_service::FullClient as CFullClient;
 
 /// `DomainInstanceStarter` used to start a domain instance node based on the given
