+++ crates/subspace-node/src/commands/run/domain.rs
@@ -2,7 +2,8 @@ use crate::commands::run::shared::RpcOptions;
 use crate::commands::shared::{store_key_in_keystore, KeystoreOptions};
 use crate::Error;
 use clap::Parser;
-use domain_client_operator::{BootstrapResult, OperatorStreams};
+use domain_client_operator::snap_sync::ConsensusChainSyncParams;
+use domain_client_operator::{BootstrapResult, ConsensusChainSyncOracleWrapper, OperatorStreams};
 use domain_eth_service::provider::EthProvider;
 use domain_eth_service::DefaultEthConfig;
 use domain_runtime_primitives::opaque::Block as DomainBlock;
@@ -36,7 +37,6 @@ use std::net::SocketAddr;
 use std::sync::Arc;
 use subspace_runtime::RuntimeApi as CRuntimeApi;
 use subspace_runtime_primitives::opaque::Block as CBlock;
-use subspace_service::domains::ConsensusChainSyncParams;
 use subspace_service::FullClient as CFullClient;
 use tokio::sync::broadcast::Receiver;
 use tokio_stream::wrappers::errors::BroadcastStreamRecvError;
